+++ functions.h
@@ -3,4 +3,4 @@ int add(int a, int b);
 int multiply(int a ,int b);
 int subtract(int a ,int b);
 int divide(int a ,int b);
-int modulus(float a ,float b);
+float modulus(int a ,int b);
